requirements.txt
Flask==2.3.2
Pillow==11.1.0
pandas>=2.2.0
openpyxl==3.1.2
pytesseract